+++ chapter4/newton-solver.py
@@ -13,10 +13,10 @@
 #     name: python3
 # ---
 
-# # Newton's method
+# # Custom Newton solvers
 # Author: Jørgen S. Dokken
 #
-# Newtons method, as used in [Non-linear Poisson](./../chapter2/nonlinpoisson_code) is a way of solving a non-linear equation as a sequence of linear equations.
+# Newtons method, as used in the [non-linear Poisson](./../chapter2/nonlinpoisson_code) problem, is a way of solving a non-linear equation as a sequence of linear equations.
 #
 # Given a function $F:\mathbb{R}^M\mapsto \mathbb{R}^M$, we have that $u_k, u_{k+1}\in \mathbb{R}^M$ is related as:
 # $$x_{k+1} = x_{k} - J_F(x_k)^{-1} F(x_k)$$
@@ -30,12 +30,13 @@
 # ## Problem specification
 # We start by importing all packages needed to solve the problem.
 
+import dolfinx
+import matplotlib.pyplot as plt
+import numpy as np
+import pyvista
+import ufl
 from mpi4py import MPI
 from petsc4py import PETSc
-import dolfinx
-import ufl
-import numpy as np
-import matplotlib.pyplot as plt
 
 
 # We will consider the following non-linear problem:
@@ -48,8 +49,10 @@ import matplotlib.pyplot as plt
 def root_0(x):
     return 3 + x[0]
 
+
 def root_1(x):
     return -1 - x[0]
+
 
 N = 10
 roots = [root_0, root_1]
@@ -68,7 +71,7 @@ uh = dolfinx.fem.Function(V)
 
 v = ufl.TestFunction(V)
 x = ufl.SpatialCoordinate(mesh)
-F = uh**2 * v * ufl.dx - 2 * uh * v * ufl.dx - (x[0]**2 + 4*x[0] + 3)* v * ufl.dx
+F = uh**2 * v * ufl.dx - 2 * uh * v * ufl.dx - (x[0]**2 + 4 * x[0] + 3) * v * ufl.dx
 residual = dolfinx.fem.form(F)
 
 # Next, we can define the jacobian $J_F$, by using `ufl.derivative`.
@@ -117,7 +120,7 @@ while i < max_iterations:
     dx.x.scatter_forward()
     # Update u_{i+1} = u_i + delta x_i
     uh.x.array[:] += dx.x.array
-    i+=1
+    i += 1
 
     # Compute norm of update
     correction_norm = dx.vector.norm(0)
@@ -136,7 +139,7 @@ print(f"Final residual {L.norm(0)}")
 
 # +
 # Plot solution for each of the iterations
-fig = plt.figure(figsize=(15,8))
+fig = plt.figure(figsize=(15, 8))
 for j, solution in enumerate(solutions[:i]):
     plt.plot(coords[sort_order], solution[sort_order], label=f"Iteration {j}")
 
@@ -148,8 +151,8 @@ for j, root in enumerate(roots):
     global_L2 = mesh.comm.allreduce(dolfinx.fem.assemble_scalar(L2_error), op=MPI.SUM)
     print(f"L2-error (root {j}) {np.sqrt(global_L2)}")
 
-    kwargs = {} if j==0 else {"label": "u_exact"}
-    plt.plot(x_spacing, root(x_spacing.reshape(1,-1)), *args, **kwargs)
+    kwargs = {} if j == 0 else {"label": "u_exact"}
+    plt.plot(x_spacing, root(x_spacing.reshape(1, -1)), *args, **kwargs)
 plt.grid()
 plt.legend()
 plt.show()
@@ -159,31 +162,139 @@ plt.show()
 # # Newton's method with DirichletBC
 # In the previous example, we did not consider handling of Dirichlet boundary conditions. 
 # For this example, we will consider the [non-linear Poisson](./../chapter2/nonlinpoisson)-problem.
+# We start by defining the mesh, the analytical solution and the forcing term $f$.
 
 # +
 def q(u):
     return 1 + u**2
 
+
 domain = dolfinx.mesh.create_unit_square(MPI.COMM_WORLD, 10, 10)
 x = ufl.SpatialCoordinate(domain)
-u_ufl = 1 + x[0] + 2*x[1]
-f = - ufl.div(q(u_ufl)*ufl.grad(u_ufl))
+u_ufl = 1 + x[0] + 2 * x[1]
+f = - ufl.div(q(u_ufl) * ufl.grad(u_ufl))
 
+
+def u_exact(x):
+    return eval(str(u_ufl))
+
+
+# -
+
+# Next, we define the boundary condition `bc`, the residual `F` and the Jacobian `J`.
+
+# +
 V = dolfinx.fem.FunctionSpace(domain, ("CG", 1))
-u_exact = lambda x: eval(str(u_ufl))
 u_D = dolfinx.fem.Function(V)
 u_D.interpolate(u_exact)
 fdim = domain.topology.dim - 1
-domain.topology.create_connectivity(fdim, fdim+1)
+domain.topology.create_connectivity(fdim, fdim + 1)
 boundary_facets = dolfinx.mesh.exterior_facet_indices(domain.topology)
 bc = dolfinx.fem.dirichletbc(u_D, dolfinx.fem.locate_dofs_topological(V, fdim, boundary_facets))
 
 uh = dolfinx.fem.Function(V)
 v = ufl.TestFunction(V)
-F = q(uh)*ufl.dot(ufl.grad(uh), ufl.grad(v))*ufl.dx - f*v*ufl.dx
+F = q(uh) * ufl.dot(ufl.grad(uh), ufl.grad(v)) * ufl.dx - f * v * ufl.dx
 J = ufl.derivative(F, uh)
 residual = dolfinx.fem.form(F)
 jacobian = dolfinx.fem.form(J)
+
 # -
+# Next, we define the matrix `A`, right hand side vector `L` and the correction function `dx`
+
+dx = dolfinx.fem.Function(V)
+A = dolfinx.fem.petsc.create_matrix(jacobian)
+L = dolfinx.fem.petsc.create_vector(residual)
+solver = PETSc.KSP().create(mesh.comm)
+solver.setOperators(A)
+
+# As we for this problem has strong Dirichlet conditions, we need to apply lifting to the right hand side of our Newton problem.
+# We previously had that we wanted to solve the system:
+# $$\begin{align}
+# J_F(x_k)\delta x_k &= - F(x_k)\\
+# x_{k+1} &= x_k + \delta x_k
+# \end{align}$$
+# we want $x_{k+1}\vert_{bc}= u_D$. However, we do not know if $x_k\vert_{bc}=u_D$.
+# Therefore, we want to apply the following boundary condition for our correction $\delta x_k$
+# $$\begin{align}
+# \delta x_k\vert_{bc} = u_D-x_k\vert_{bc}
+# \end{align}$$
+# We therefore arrive at the following Newton scheme
+
+
+i = 0
+error = dolfinx.fem.form(ufl.inner(uh - u_ufl, uh - u_ufl) * ufl.dx(metadata={"quadrature_degree": 4}))
+L2_error = []
+dx_norm = []
+while i < max_iterations:
+    # Assemble Jacobian and residual
+    L.zeroEntries()
+    A.zeroEntries()
+    dolfinx.fem.petsc.assemble_matrix(A, jacobian, bcs=[bc])
+    A.assemble()
+    dolfinx.fem.petsc.assemble_vector(L, residual)
+    L.scale(-1)
+    # Compute b - J(u_D-u_(i-1))
+    dolfinx.fem.petsc.apply_lifting(L, [jacobian], [[bc]], x0=[uh.vector], scale=1) 
+    L.ghostUpdate(addv=PETSc.InsertMode.ADD, mode=PETSc.ScatterMode.REVERSE)
+    # Set dx|_bc = u_{i-1}-u_D
+    dolfinx.fem.petsc.set_bc(L, [bc], uh.vector, 1.0)
+    L.ghostUpdate(addv=PETSc.InsertMode.INSERT_VALUES, mode=PETSc.ScatterMode.FORWARD)
+
+    # Solve linear problem
+    solver.solve(L, dx.vector)
+    dx.x.scatter_forward()
+
+    # Update u_{i+1} = u_i + delta x_i
+    uh.x.array[:] += dx.x.array
+    i += 1
+
+    # Compute norm of update
+    correction_norm = dx.vector.norm(0)
+
+    # Compute L2 error comparing to the analytical solution
+    L2_error.append(np.sqrt(mesh.comm.allreduce(dolfinx.fem.assemble_scalar(error), op=MPI.SUM)))
+    dx_norm.append(correction_norm)
+    
+    print(f"Iteration {i}: Correction norm {correction_norm}, L2 error: {L2_error[-1]}")
+    if correction_norm < 1e-10:
+        break
+
+# We plot the $L^2$-error and the residual norm ($\delta x$) per iteration
+
+fig = plt.figure(figsize=(15, 8))
+plt.subplot(121)
+plt.plot(np.arange(i), L2_error)
+plt.title(r"$L^2(\Omega)$-error of $u_h$")
+ax = plt.gca()
+ax.set_yscale('log')
+plt.xlabel("Iterations")
+plt.ylabel(r"$L^2$-error")
+plt.grid()
+plt.subplot(122)
+plt.title(r"Residual of $\vert\vert\delta x_i\vert\vert$") 
+plt.plot(np.arange(i), dx_norm)
+ax = plt.gca()
+ax.set_yscale('log')
+plt.xlabel("Iterations")
+plt.ylabel(r"$\vert\vert \delta x\vert\vert$")
+plt.grid()
+
+# We compute the max error and plot the solution
+
+error_max = domain.comm.allreduce(np.max(np.abs(uh.x.array - u_D.x.array)), op=MPI.MAX)
+if domain.comm.rank == 0:
+    print(f"Error_max: {error_max:.2e}")
+
+pyvista.set_jupyter_backend("pythreejs")
+u_topology, u_cell_types, u_geometry = dolfinx.plot.create_vtk_mesh(V)
+u_grid = pyvista.UnstructuredGrid(u_topology, u_cell_types, u_geometry)
+u_grid.point_data["u"] = uh.x.array.real
+u_grid.set_active_scalars("u")
+u_plotter = pyvista.Plotter()
+u_plotter.add_mesh(u_grid, show_edges=True)
+u_plotter.view_xy()
+if not pyvista.OFF_SCREEN:
+    u_plotter.show()